Setup volley and make api call

+++ app/src/main/java/co/bildit/addin/weatheria/model/Main.java
@@ -33,7 +33,7 @@ private Integer grndLevel;
 private Integer humidity;
 @SerializedName("temp_kf")
 @Expose
-private Integer tempKf;
+private Double tempKf;
 
 public Double getTemp() {
 return temp;
@@ -99,11 +99,11 @@ public void setHumidity(Integer humidity) {
 this.humidity = humidity;
 }
 
-public Integer getTempKf() {
+public Double getTempKf() {
 return tempKf;
 }
 
-public void setTempKf(Integer tempKf) {
+public void setTempKf(Double tempKf) {
 this.tempKf = tempKf;
 }
